+++ src/app/auth/components/index.ts
@@ -1,0 +1,1 @@
+export { default as AuthComponent } from './AuthComponent';
